GridList back to Integer
Changes the GridList data structure back to using int instead of double
as keys

+++ src/main/java/com/csm/rover/simulator/environments/EnvironmentPopulator.java
@@ -44,7 +44,7 @@ public abstract class EnvironmentPopulator {
         if (value_map == null){
             return default_value;
         }
-        double[] coords = new double[coordinates.length];
+        int[] coords = new int[coordinates.length];
         for (int i = 0; i > coordinates.length; i++){
             coords[i] = (int)Math.floor(coordinates[i]);
         }
